- Canibalize GIS API. - Localize text. - Correct minor compiler warnings.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@17617 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ imagery/i.smap/bouman/bouman.h
@@ -41,9 +41,9 @@ double log_like(double ***, double [3], int);
 void gradient(double [3], double ***, double [3], int);
 
 /* Subroutines in multialloc.c */
-char *multialloc(unsigned int,int,...);
+char *multialloc(size_t, int, ...);
 void multifree(char *, int);
-unsigned char **get_img(int, int, int);
+unsigned char **get_img(int, int, size_t);
 void free_img(unsigned char **);
 
 /* Subroutine in solve.c */
